sagemaker/requirements.txt
boto3>=1.34.0
sagemaker>=2.200.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
